fix numeric type

+++ cpp/src/arrow/engine/substrait/serde.cc
@@ -74,12 +74,12 @@ static DeclarationFactory MakeConsumingSinkDeclarationFactory(
 
 static compute::Declaration ProjectByNamesDeclaration(compute::Declaration input,
                                                       std::vector<std::string> names) {
-  int names_size = names.size();
+  size_t names_size = names.size();
   if (names.size() == 0) {
     return input;
   }
   std::vector<compute::Expression> expressions;
-  for (int i = 0; i < names_size; i++) {
+  for (size_t i = 0; i < names_size; i++) {
     expressions.push_back(compute::field_ref(FieldRef(i)));
   }
   return compute::Declaration::Sequence(
